+++ package.json
@@ -13,6 +13,7 @@
     "@types/react-router-dom": "^5.1.3",
     "bootstrap": "^4.4.1",
     "react": "^16.12.0",
+    "react-bootstrap": "^1.0.0-beta.16",
     "react-dom": "^16.12.0",
     "react-router-dom": "^5.1.2",
     "react-scripts": "3.4.0",
