+++ src/styles/index.css
@@ -1,3 +1,4 @@
+/*Leftover css*/
 body {
   margin: 0;
   font-family: -apple-system, BlinkMacSystemFont, 'Segoe UI', 'Roboto', 'Oxygen',
